+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ dev
    - Fully implement a new backend module
       - Allow NumPy and Torch backends
 - Coding style
+   - Run tests automatically for NumPy and Torch
    - Many unifications due to the added backend support
    - Stricter input types
    - Ignore Ruffs PLC0415 rule